requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
Flask_MySQLdb==1.0.1
Jinja2==3.1.2
numpy==1.23.4
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.7.0.72
PyJWT==2.7.0
PyJWT==2.7.0
Werkzeug==2.2.3
